requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
jupyterlab>=4.0.0
kaggle>=1.5.0
